api/requirements.txt
flask
flask-cors
google-cloud-storage
google-auth-oauthlib
pandas